Create README.md
Hi! Am Collins Stevards 19 years of age, a self-taught programmer, web design and developer, and finally a game developer. 
Am a beginner at all this level and am willing to move forward to do my possible best, to become the best of all the best programmers out there.
And for sure I will be happy to meet you for help and some one-on-one challenges all though am a beginner but am proud to be challenged and that is my favourite part of it. And am proud to join you guys very very excited.